pyproject.toml
[project]
name = "dagster_pipeline"
description = "Common Ground data pipeline — Dagster + dlt + DuckLake"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster==1.12.19",
    "dagster-dlt>=0.28.0",
    "dlt[duckdb,filesystem]>=1.23.0",
    "duckdb==1.5.0",  # pinned until duckpgq/lsh/anofox_forecast rebuild for 1.5.1
    "ducklake>=0.1.0",
    "httpx>=0.27",
    "pyarrow>=18.0",
    "tenacity>=9.0",
    "s3fs>=2024.0",
    "psutil>=5.9",
    "orjson>=3.11.7",
    "dagster-docker>=0.28.0",
    "splink>=4.0",
    "lancedb>=0.30.1",
    "google-genai>=1.69.0",
    "anthropic>=0.86.0",
    "pylance>=4.0.0",
]

[project.optional-dependencies]
ml = [
    "blockingpy>=0.1.0",
    "semhash>=0.1.0",
    "bertopic>=0.16.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }
packages = ["src/dagster_pipeline"]

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dagster_pipeline"
registry_modules = [
    "dagster_pipeline.components.*",
]
